Todas las definiciones en nuevo archivo. Pendiente depurar programa
